Split code block logic into multiple modules
