+++ src_js/components/main_content/enhanced_code_blocks/parseCodeHighlightRanges.ts
@@ -1,0 +1,57 @@
+/**
+ * Parse a string reprenting a list of line numbers, some of which may be
+ * ranges. The parsed output is a Set of line numbers that are included in the
+ * range.
+ *
+ * For instance, the string `'13, 24-26, 25-27'` is parsed as
+ * `Set([13, 24, 25, 26, 27])`
+ *
+ * @param rawHighlightRanges A comma-separated string representing ranges
+ * @param maxLineNumber The maximum valid line number
+ */
+export function parseCodeHighlightRanges(
+  rawHighlightRanges: string | null,
+  maxLineNumber: number,
+): Set<number> {
+  const highlightedLines = new Set<number>();
+  if (!rawHighlightRanges) {
+    return highlightedLines;
+  }
+
+  const ranges = rawHighlightRanges.split(',');
+  ranges.forEach((range) => {
+    // First check if it's a single number
+    const potentialLineNum = +range;
+    if (isNumWithinInclusiveRange(potentialLineNum, 1, maxLineNumber)) {
+      highlightedLines.add(potentialLineNum);
+    } else {
+      const rangeParts = range.trim().split('-');
+      if (rangeParts.length === 2) {
+        const lower = +rangeParts[0];
+        const upper = +rangeParts[1];
+        if (
+          isNumWithinInclusiveRange(lower, 1, maxLineNumber) &&
+          isNumWithinInclusiveRange(upper, 1, maxLineNumber) &&
+          lower <= upper
+        ) {
+          for (let i = lower; i <= upper; ++i) {
+            highlightedLines.add(i);
+          }
+        }
+      }
+    }
+  });
+  return highlightedLines;
+}
+
+/**
+ * Return a boolean indicating whether `num` is in the range [`lower`, `upper`]
+ * (inclusive).
+ */
+function isNumWithinInclusiveRange(
+  num: number | null,
+  lower: number,
+  upper: number,
+): boolean {
+  return num != null && !Number.isNaN(num) && num >= lower && num <= upper;
+}
