--- src_js/components/main_content/types.ts
@@ -1,5 +1,0 @@
-export enum CodeblockVariant {
-  ENHANCED = 'enhanced',
-  NO_LINE_NUMBERS = 'no-line-numbers',
-  LEGACY = 'legacy',
-}
